requirements.txt
discord.py[voice]==1.0.1
future
